splited Data project into Data, Core and Services projects

+++ src/Applicanty.API/Controllers/CandidateController.cs
@@ -1,6 +1,6 @@
 ﻿using Applicanty.API.Models.Response;
-using Applicanty.Data.Entity;
-using Applicanty.Data.Services;
+using Applicanty.Core.Model;
+using Applicanty.Services.Abstract;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Linq;
@@ -19,7 +19,7 @@ namespace Applicanty.API.Controllers
         }
 
         [HttpGet("{id}")]
-        public IActionResult Get(long id)
+        public IActionResult Get(int id)
         {
             try
             {
@@ -94,7 +94,7 @@ namespace Applicanty.API.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Archive(long id)
+        public IActionResult Archive(int id)
         {
             try
             {
